V1.2.1 : Registration and Authentication
Bug Fixes-Phone Number

+++ src/main/java/com/zirconiatic/employeemanager/dao/RegistrationRepository.java
@@ -1,5 +1,7 @@
 package com.zirconiatic.employeemanager.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -10,7 +12,7 @@ import com.zirconiatic.employeemanager.entity.Registration;
 
 @Repository
 @Transactional
-public interface RegistrationRepository extends JpaRepository<Registration, Integer>{
+public interface RegistrationRepository extends JpaRepository<Registration, Long>{
 
 	//Using Native Query
 	@Query(nativeQuery = true,value = "select * from registration where email_id=:eid and password=:pass")
